// frontend/src/types/auth.ts
export interface User {
  _id: string;
  username: string;
  email: string;
  isAdmin: boolean;
  token: string;
}

export interface AuthState {
  user: User | null;
  loading: boolean;
  error: string | null;
}

export interface LoginCredentials {
  email: string;
  password: string;
}

export interface RegisterData {
  username: string;
  email: string;
  password: string;
}